final_project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "social-listening"
version = "0.1.0"
dependencies = [
    "tweepy",
    "Mastodon.py",
    "sudachipy",
    "sudachidict-full",
    "transformers",
    "torch",
    "pandas",
    "streamlit",
    "python-dotenv",
    "weasyprint",
    "markdown2",
    "emoji",
]
requires-python = ">=3.11"
